Working on Texture stuff Added grass.png Minor tweaks to code Fixed oninoni's Memory Leak fail with tiles Added VisualLeakDetector (look in ts browser, there you can download and install it into your visual studio) gvec2 template with generic type of elements for - vec2 = floats - ivec2 = ints - xvec2 = cardinals (= unsigned int) If you actually read all this, oninoni, look in the KeyManager.cpp (I would have just called it "Controls" but I know you like your manager >.<) there is a little comment for you

+++ ProjectPewPew/ProjectPewPew/Game.h
@@ -12,9 +12,9 @@ private:
 
     Player* player;
 
-	double oldTime;
-    double deltaTime;
-    double runTime;
+	float oldTime;
+    float deltaTime;
+    float runTime;
 
     Shader* testShader;
 
@@ -34,7 +34,7 @@ public:
     Game(int* argc, char** argv);
     ~Game();
 
-    double getDeltaTime();
-    double getRunTime();
+    float getDeltaTime();
+    float getRunTime();
 };
